아이템 기능 및 레벨 전환 기능 구현.

+++ Source/PawnClass/Public/JPotionItem.h
@@ -13,7 +13,7 @@ public:
 	AJPotionItem();
 
 	UPROPERTY(EditAnywhere, BlueprintReadWrite, Category = "Item")
-	int32 HealAmount;
+	float HealAmount;
 	
 	virtual void ActivateItem(AActor* Activator) override;
 };
